+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 0.2.6
 -----
 - Fixed `HBase::Scoped#filter` method to take short integer (byte, short, int) values
+- Fixed `HBase::Scoped#range` method to take short integer (byte, short, int) values
 
 0.2.5
 -----